# mascotas-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Mascotas y Personas",
        "description": "Documentación de la API RESTful con autenticación JWT",
        "version": "1.0.0"
    },
    "paths": {
        "/api/mascotas": {
            "get": {
                "tags": [
                    "Mascotas"
                ],
                "summary": "Listar mascotas paginadas",
                "operationId": "b06bcf95c18bee8ed15647d8fa088cbf",
                "responses": {
                    "200": {
                        "description": "Listado de mascotas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Mascotas"
                ],
                "summary": "Crear una nueva mascota",
                "operationId": "7ef45de4bf722d348a63d7f462b44898",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "especie",
                                    "raza",
                                    "edad",
                                    "persona_id"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Toby"
                                    },
                                    "especie": {
                                        "type": "string",
                                        "example": "perro"
                                    },
                                    "raza": {
                                        "type": "string",
                                        "example": "Labrador"
                                    },
                                    "edad": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "persona_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mascota creada"
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/mascotas/{id}": {
            "get": {
                "tags": [
                    "Mascotas"
                ],
                "summary": "Obtener una mascota por ID",
                "operationId": "1a2eaf8302ca37018e3c8449c1b7c9ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la mascota",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos de la mascota"
                    },
                    "404": {
                        "description": "Mascota no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Mascotas"
                ],
                "summary": "Actualizar mascota",
                "operationId": "58258f41d5e256a91ee928955dbd1ad8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la mascota a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Toby"
                                    },
                                    "edad": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mascota actualizada"
                    },
                    "404": {
                        "description": "Mascota no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Mascotas"
                ],
                "summary": "Eliminar mascota",
                "operationId": "ff434682a0ddbfa68e8fc15a98d51de7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la mascota",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Mascota eliminada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/razas": {
            "get": {
                "tags": [
                    "Mascotas"
                ],
                "summary": "Obtener razas desde API externa",
                "operationId": "9809d4491decf1694ccbd93d5225e6ce",
                "parameters": [
                    {
                        "name": "especie",
                        "in": "query",
                        "description": "perro o gato",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "perro",
                                "gato"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de razas"
                    },
                    "422": {
                        "description": "Parámetro inválido"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}